fci-platform/requirements.txt
fastapi[standard]==0.135.1
uvicorn[standard]==0.41.0
anthropic==0.84.0
pymongo==4.16.0
pydantic==2.12.5
pydantic-settings>=2.7.0
pyyaml>=6.0.2
python-dotenv>=1.0.1
sse-starlette==3.2.0
